emoteam-functions/requirements.txt
audioread==3.0.1
azure-core==1.30.1
azure-functions==1.18.0
azure-storage-blob==12.19.1
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
contourpy==1.2.1
cryptography==42.0.5
cycler==0.12.1
decorator==5.1.1
filelock==3.13.3
fonttools==4.51.0
fsspec==2024.3.1
idna==3.6
imageio==2.34.0
isodate==0.6.1
Jinja2==3.1.3
joblib==1.4.0
kiwisolver==1.4.5
lazy_loader==0.4
liac-arff==2.5.0
librosa==0.10.1
llvmlite==0.42.0
MarkupSafe==2.1.5
matplotlib==3.8.4
mpmath==1.3.0
msgpack==1.0.8
networkx==3.3
numba==0.59.1
numpy==1.26.4
packaging==24.0
pandas==2.2.1
pillow==10.3.0
platformdirs==4.2.0
pooch==1.8.1
pycparser==2.22
pydub==0.25.1
pyparsing==3.1.2
python-dateutil==2.9.0.post0
pytz==2024.1
requests==2.31.0
scikit-learn==1.4.2
scipy==1.13.0
six==1.16.0
soundfile==0.12.1
soxr==0.3.7
sympy==1.12
threadpoolctl==3.4.0
torch @ https://download.pytorch.org/whl/cpu-cxx11-abi/torch-2.2.0%2Bcpu.cxx11.abi-cp311-cp311-linux_x86_64.whl
typing_extensions==4.11.0
tzdata==2024.1
urllib3==2.2.1
